Fazendo mudancas
