+++ bin/.gitignore
@@ -1,0 +1,8 @@
+/Bucket.class
+/IndiceHash.class
+/InterfaceGrafica$1.class
+/InterfaceGrafica.class
+/Main.class
+/Pagina.class
+/Tabela.class
+/Tupla.class
