+++ src/Main.java
@@ -4,7 +4,8 @@ public class Main{
 	public static void main(String[] args) {
 		//Dados do Usuario
 		String chave = "ZZZ";
-		IndiceHash indice = new IndiceHash("C:\\Users\\Luan\\Documents\\Banco de dados\\words.txt");
+		Tabela indice = new Tabela("C:\\Users\\User\\Downloads\\words.txt");
+		//Especificar caminho do arquivo na máquina
 		try {
 			System.out.println(indice.chaveBusca(chave));
 			System.out.println(indice.tableScan(chave));
